Rename the basic access controller to htpasswd
