+++ cmd/registry/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/docker/distribution/configuration"
 	"github.com/docker/distribution/context"
 	_ "github.com/docker/distribution/health"
-	_ "github.com/docker/distribution/registry/auth/basic"
+	_ "github.com/docker/distribution/registry/auth/htpasswd"
 	_ "github.com/docker/distribution/registry/auth/silly"
 	_ "github.com/docker/distribution/registry/auth/token"
 	"github.com/docker/distribution/registry/handlers"
