+++ registry/auth/htpasswd/access_test.go
@@ -1,4 +1,4 @@
-package basic
+package htpasswd
 
 import (
 	"io/ioutil"
