+++ registry/auth/htpasswd/htpasswd_test.go
@@ -1,4 +1,4 @@
-package basic
+package htpasswd
 
 import (
 	"fmt"
